Move files and update headers to respect JE standards. Fix all warnings.
