+++ Framework/Sources/MockHTTPMethod.swift
@@ -1,0 +1,12 @@
+//  MockHTTPMethod.swift
+
+import Foundation
+
+public enum MockHTTPMethod: String {
+    case get     = "GET"
+    case head    = "HEAD"
+    case post    = "POST"
+    case put     = "PUT"
+    case patch   = "PATCH"
+    case delete  = "DELETE"
+}
