--- Shock/Classes/MockHTTPMethod.swift
@@ -1,18 +1,0 @@
-//
-//  MockHTTPMethod.swift
-//  Shock
-//
-//  Created by Jack Newcombe on 27/06/2018.
-//  Copyright © 2018 Just Eat. All rights reserved.
-//
-
-import Foundation
-
-public enum MockHTTPMethod: String {
-    case get     = "GET"
-    case head    = "HEAD"
-    case post    = "POST"
-    case put     = "PUT"
-    case patch   = "PATCH"
-    case delete  = "DELETE"
-}
