+++ Tests/Sources/ApiCallRequestDataTests.swift
@@ -1,0 +1,35 @@
+//  ApiCallRequestData.swift
+
+import XCTest
+@testable import Shock
+
+class ApiCallRequestDataTests: XCTestCase {
+    
+    var server: MockServer!
+    
+    override func setUp() {
+        super.setUp()
+        server = MockServer(portRange: 9090...9099, bundle: Bundle.module)
+        server.shouldSendNotFoundForMissingRoutes = true
+        server.start()
+    }
+    
+    override func tearDown() {
+        server.stop()
+        super.tearDown()
+    }
+    
+    func testGetApiCallRequestData() {
+        let route: MockHTTPRoute = .simple(method: .get, urlPath: "/simple", code: 200, filename: "testSimpleRoute.txt")
+        server.setup(route: route)
+        
+        let expectation = self.expectation(description: "Expect 200 response with response body")
+        
+        HTTPClient.get(url: "\(server.hostURL)/simple") { code, body, headers, error in
+            XCTAssertEqual(code, 200)
+            XCTAssertEqual(body, "testSimpleRoute test fixture\n")
+            expectation.fulfill()
+        }
+        self.waitForExpectations(timeout: 2.0, handler: nil)
+    }
+}
