+++ Tests/Sources/Resources/Fixtures/testSimpleRouteWithVariables.txt
