+++ Tests/Sources/Resources/Fixtures/testTemplatedRoute.mustache
